# [crypto] Switch to strict Ed25519 signature verification