--- a/crypto/crypto/Cargo.toml
+++ b/crypto/crypto/Cargo.toml
@@ -66,3 +66,7 @@ harness = false
 [[bench]]
 name = "noise"
 harness = false
+
+[[bench]]
+name = "ed25519"
+harness = false