This adopts the strict check on the order of the R component of an Ed25519 signature, as this project used to have, but removed for performance reasons in #876 . This check is routinely used in LibSodium, and hinted at in [RFC8032 5.1.7 §3](https://tools.ietf.org/html/rfc8032#section-5.1.7).

This check is discussed in [[Brendel at al. 2020]](https://eprint.iacr.org/2020/823) §5.4 and provides message binding, along with protection against substitution attacks. A proptest added in the present PR makes a possible substitution explicit, and should make it clear that this check will probably not reject any "clumsy but honest" signatures or pose significant backwards compatibility issues.

Finally, a minimalist benchmark is added to check that https://github.com/dalek-cryptography/ed25519-dalek/pull/98 provides a performant way to do this check.

Before:

```
Ed25519 signature verification                                                                                                                                                                                                                                                                        
                        time:   [48.504 us 48.974 us 49.757 us]                                                                                                                                                                                                                                       
Found 14 outliers among 100 measurements (14.00%)                                                                                                                                                                                                                                                     
  1 (1.00%) high mild                                                                                                                                                                                                                                                                                 
  13 (13.00%) high severe                                                                                                                                                                                                                                                                             
```

After:

```
Ed25519 signature verification                                                                              
                        time:   [51.086 us 51.522 us 52.113 us]
Found 27 outliers among 100 measurements (27.00%)
  7 (7.00%) low mild
  5 (5.00%) high mild
  15 (15.00%) high severe
```